+++ android/sdk/src/main/java/com/tencent/mtt/hippy/bridge/HippyBridgeImpl.java
@@ -52,6 +52,7 @@ public class HippyBridgeImpl implements HippyBridge, DevRemoteDebugProxy.OnRecei
 	private BridgeCallback						mBridgeCallback;
 	private boolean								mInit						= false;
 	private boolean								mIsDevModule				= false;
+	private String                              mDebugServerHost;
 	private boolean								mSingleThreadMode			= false;
 	private boolean								mBridgeParamJson;
 	private HippyBuffer                         mHippyBuffer;
@@ -59,12 +60,14 @@ public class HippyBridgeImpl implements HippyBridge, DevRemoteDebugProxy.OnRecei
 	private String                              mDebugGobalConfig;
 	private NativeCallback                      mDebugInitJSFrameworkCallback;
 
-	public HippyBridgeImpl(Context context, BridgeCallback callback, boolean singleThreadMode, boolean jsonBrige, boolean isDevModule)
+	public HippyBridgeImpl(Context context, BridgeCallback callback, boolean singleThreadMode,
+			boolean jsonBrige, boolean isDevModule, String debugServerHost)
 	{
 		this.mBridgeCallback = callback;
 		this.mSingleThreadMode = singleThreadMode;
 		this.mBridgeParamJson = jsonBrige;
 		this.mIsDevModule = isDevModule;
+		this.mDebugServerHost = debugServerHost;
 
 		synchronized (sBridgeSyncLock)
 		{
@@ -85,7 +88,7 @@ public class HippyBridgeImpl implements HippyBridge, DevRemoteDebugProxy.OnRecei
 				this.mCodeCacheThreadExecutor.allowCoreThreadTimeOut(true);
 			}
 		}
-        
+
 		if (!mBridgeParamJson)
 		{
 			mHippyBuffer = new HippyBuffer();
@@ -109,7 +112,10 @@ public class HippyBridgeImpl implements HippyBridge, DevRemoteDebugProxy.OnRecei
 		{
 			mDebugWebSocketClient = new DebugWebSocketClient();
 			mDebugWebSocketClient.setOnReceiveDataCallback(this);
-			mDebugWebSocketClient.connect(String.format(Locale.US, "ws://%s/debugger-proxy?role=android_client", "localhost:38989"), new DebugWebSocketClient.JSDebuggerCallback()
+			if (TextUtils.isEmpty(mDebugServerHost)) {
+				mDebugServerHost = "localhost:38989";
+			}
+			mDebugWebSocketClient.connect(String.format(Locale.US, "ws://%s/debugger-proxy?role=android_client", mDebugServerHost), new DebugWebSocketClient.JSDebuggerCallback()
 			{
 				@Override
 				public void onSuccess(String response)
@@ -239,13 +245,14 @@ public class HippyBridgeImpl implements HippyBridge, DevRemoteDebugProxy.OnRecei
 				mCodeCacheThreadExecutor = null;
 			}
 		}
-		
+
 		if (!mBridgeParamJson && mHippyBuffer != null)
 		{
 			mHippyBuffer.release();
 		}
 
 		destroy(mV8RuntimeId, mSingleThreadMode, callback);
+		mV8RuntimeId = 0;
 		mBridgeCallback = null;
 	}