+++ examples/hippy-react-demo/src/components/ListView/index.jsx
@@ -44,6 +44,7 @@ const mockDataArray = [
 const styles = StyleSheet.create({
   container: {
     backgroundColor: '#fff',
+    collapsable: false,
   },
   itemContainer: {
     padding: 12,
@@ -186,6 +187,7 @@ export default class ListExample extends React.Component {
         getRowType={this.getRowType}
         getRowKey={this.getRowKey}
         initialListSize={15}
+        rowShouldSticky={index => index === 2}
       />
     );
   }