+++ examples/hippy-vue-demo/scripts/hippy-webpack.dev.js
@@ -4,8 +4,20 @@ const webpack           = require('webpack');
 const VueLoaderPlugin   = require('vue-loader/lib/plugin');
 const pkg               = require('../package.json');
 
+let cssLoader = '@hippy/vue-css-loader';
+const hippyVueCssLoaderPath = path.resolve(__dirname, '../../../packages/hippy-vue-css-loader/dist/index.js');
+if (fs.existsSync(hippyVueCssLoaderPath)) {
+  /* eslint-disable-next-line no-console */
+  console.warn(`* Using the @hippy/vue-css-loader in ${hippyVueCssLoaderPath}`);
+  cssLoader = hippyVueCssLoaderPath;
+} else {
+  /* eslint-disable-next-line no-console */
+  console.warn('* Using the @hippy/vue-css-loader defined in package.json');
+}
+
 module.exports = {
   mode: 'development',
+  devtool: 'eval-source-map',
   watch: true,
   watchOptions: {
     aggregateTimeout: 1500,
@@ -35,6 +47,7 @@ module.exports = {
         test: /\.vue$/,
         use: [
           'vue-loader',
+          'scope-loader',
           'unicode-loader',
         ],
       },
@@ -64,7 +77,14 @@ module.exports = {
       {
         test: /\.css$/,
         use: [
-          '@hippy/vue-css-loader',
+          cssLoader,
+        ],
+      },
+      {
+        test: /\.less$/,
+        use: [
+          cssLoader,
+          'less-loader',
         ],
       },
       {