+++ ios/sdk/module/uimanager/HippyUIManager.mm
@@ -1328,6 +1328,32 @@ HIPPY_EXPORT_METHOD(measureInWindow:(nonnull NSNumber *)hippyTag
             callback(@[]);
             return;
         }
+        UIView *rootView = viewRegistry[view.rootTag];
+        if (!rootView) {
+            HippyLogWarn(@"measure cannot find view's root view with tag #%@", hippyTag);
+            callback(@[]);
+            return;
+        }
+        
+        CGRect windowFrame = [rootView convertRect:view.frame fromView:view.superview];
+        callback(@[@{@"width":@(CGRectGetWidth(windowFrame)),
+                     @"height": @(CGRectGetHeight(windowFrame)),
+                     @"x":@(windowFrame.origin.x),
+                     @"y":@(windowFrame.origin.y)}]);
+    }];
+}
+
+HIPPY_EXPORT_METHOD(measureInAppWindow:(nonnull NSNumber *)hippyTag
+                callback:(HippyResponseSenderBlock)callback)
+{
+    [self addUIBlock:^(__unused HippyUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
+        UIView *view = viewRegistry[hippyTag];
+        if (!view) {
+            // this view was probably collapsed out
+            HippyLogWarn(@"measure cannot find view with tag #%@", hippyTag);
+            callback(@[]);
+            return;
+        }
                 
         CGRect windowFrame = [view.window convertRect:view.frame fromView:view.superview];
         callback(@[@{@"width":@(CGRectGetWidth(windowFrame)),