+++ packages/hippy-react-web/src/components/text-input.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable no-unneeded-ternary */
+
 import React from 'react';
 import { formatWebStyle } from '../adapters/transfer';
 import applyLayout from '../adapters/apply-layout';
@@ -49,7 +51,7 @@ export function TextInput(props_) {
     }
   }
 
-  const { style, keyboardType } = props;
+  const { style, keyboardType, editable = true } = props;
   let inputType = 'text';
   if (keyboardType) {
     if (keyboardType === 'numeric' || keyboardType === 'phone-pad') {
@@ -63,6 +65,7 @@ export function TextInput(props_) {
   const newProps = Object.assign({}, props, {
     style: formatWebStyle(style),
     type: inputType,
+    readOnly: editable ? false : true,
   });
 
   if (typeof newProps.onChangeText === 'function') {
@@ -75,8 +78,11 @@ export function TextInput(props_) {
   }
   delete newProps.keyboardType;
   delete newProps.onLayout;
+  delete newProps.editable;
+
+  const { multiline, ..._newProps } = newProps;
   return (
-    <input {...newProps} />
+    multiline ? <textarea cols={20} rows={2} {..._newProps} /> : <input {..._newProps} />
   );
 }