+++ packages/hippy-react/src/components/refresh-wrapper.tsx
@@ -14,6 +14,8 @@ interface RefreshWrapperItemViewProps {
 
 /**
  * Simply to implement the drag down to refresh feature.
+ *
+ * @deprecated
  * @noInheritDoc
  */
 class RefreshWrapper extends React.Component<RefreshWrapperProps, {}> {